policies

+++ app/Http/Controllers/AboutUsController.php
@@ -29,6 +29,7 @@ class AboutUsController extends Controller
      */
     public function create()
     {
+        $this->authorize("create", AboutUs::class);
         return view('backoffice.pages.aboutUs.aboutUsCreate');
     }
 
@@ -40,6 +41,7 @@ class AboutUsController extends Controller
      */
     public function store(Request $request)
     {
+        // $this->authorize("create", AboutUs::class);
         $request->validate([
 
             "emoji" => "required",
@@ -78,6 +80,7 @@ class AboutUsController extends Controller
      */
     public function edit(AboutUs $about)
     {
+        $this->authorize("edit", AboutUs::class);
         return view('backoffice.pages.aboutUs.aboutUsEdit', compact('about'));
     }
 
@@ -90,6 +93,7 @@ class AboutUsController extends Controller
      */
     public function update(Request $request, AboutUs $about)
     {
+        $this->authorize("update",AboutUs::class);
         $request->validate([
             "emoji"=> "required",
             "titre"=> "required",
@@ -112,6 +116,7 @@ class AboutUsController extends Controller
      */
     public function destroy(AboutUs $about)
     {   
+        $this->authorize("delete", AboutUs::class);
         $about->delete();
         return redirect()->route('about.index')->with('message', 'supprimé avec succès');
     }
